moved check playlist selection to utils, completed browse seleciton for playlist
