+++ lib/components/Browse.js
@@ -4,30 +4,32 @@ import { getCategories, getPlaylistsFromCategory } from '@/services/spotify';
 import { STRINGS } from '@/resources';
 import { PlaylistCard } from './';
 import Carousel from 'react-native-snap-carousel';
+import { handlePlaylistSelection } from './utils';
 import { Dimensions } from "react-native";
 
 
 const Browse = ({ token, setSongs }) => {
   const [categories, setCategories] = useState([]);
   const [currentDisplay, setCurrentDisplay] = useState('categories');
-  const [playlists, setPlaylists] = useState([]);
+  const [playlists, setPlaylists] = useState(null);
 	useEffect(() => {
     token && getCategories(token, setCategories);
   }, [token]);
-  const handlePress = id => {
+  const handlePress = async id => {
     if (currentDisplay === 'categories') {
-      getPlaylistsFromCategory(token, id, setPlaylists);
+      const response = await getPlaylistsFromCategory(token, id);
+      const responseJSON = JSON.parse(response);
+      setPlaylists(responseJSON.playlists);
       setCurrentDisplay('playlists');
     } else if (currentDisplay === 'playlists') {
-
+      handlePlaylistSelection(token, id, setSongs);
     }
   };
   const renderCard = ({item, index}) => (
     <PlaylistCard
       key={index}
       id={item.id}
-      href={item.href}
-      img={item.icons[0].url}
+      img={item.image || item.icons[0].url}
       name={item.name}
       onPress={handlePress}
     />
@@ -39,7 +41,7 @@ const Browse = ({ token, setSongs }) => {
           <Text style={{fontWeight: 'bold'}}>{STRINGS.browseTitle}</Text>
         </View>
         <Carousel
-          data={categories}
+          data={playlists || categories}
           renderItem={renderCard}
           layout={'stack'}
           sliderWidth={screenWidth}
