+++ lib/components/Search.js
@@ -1,29 +1,18 @@
 import React, { useState } from 'react';
 import { View, StyleSheet, Text, TextInput, Alert } from 'react-native';
-import { searchPlaylists, getSongsFromPlaylist } from '@/services/spotify';
-import { STRINGS, constants } from '@/resources';
+import { searchPlaylists } from '@/services/spotify';
+import { STRINGS } from '@/resources';
 import { PlaylistCard } from './';
 import Carousel from 'react-native-snap-carousel';
+import { handlePlaylistSelection } from './utils';
 import { Dimensions } from "react-native";
 
-
-const isValidPlaylist = songs => {
-  return songs.tracksWithPreview >= constants.MIN_SONGS_TO_PLAY;
-};
 
 const Search = ({ token, setSongs }) => {
   const [results, setResults] = useState([]);
   const [value, onChangeText] = useState('');
-  const handlePress = async id => {
-    const response = JSON.parse(await getSongsFromPlaylist(token, id));
-    if (isValidPlaylist(response)){
-      setSongs(response)
-    } else {
-      Alert.alert(
-        `This playlist doesn't have enough playable songs... :(`,
-        'Artist(s) in this playlist may not allow samples of their music to be played',
-      );
-    }
+  const handlePress = id => {
+    handlePlaylistSelection(token, id, setSongs);
   }
   const renderCard = ({item, index}) => (
     <PlaylistCard
